chore: rename baseCommandNew to baseCommand
