+++ src/cmds/categories/create.ts
@@ -2,7 +2,7 @@ import { Args, Flags } from '@oclif/core';
 import chalk from 'chalk';
 import { Headers } from 'node-fetch';
 
-import BaseCommand from '../../lib/baseCommandNew.js';
+import BaseCommand from '../../lib/baseCommand.js';
 import { keyFlag, versionFlag } from '../../lib/flags.js';
 import getCategories from '../../lib/getCategories.js';
 import readmeAPIFetch, { cleanHeaders, handleRes } from '../../lib/readmeAPIFetch.js';
