+++ src/cmds/custompages.ts
@@ -1,6 +1,6 @@
 import { Args, Flags } from '@oclif/core';
 
-import BaseCommand from '../lib/baseCommandNew.js';
+import BaseCommand from '../lib/baseCommand.js';
 import { githubFlag, keyFlag } from '../lib/flags.js';
 import syncDocsPath from '../lib/syncDocsPath.js';
