+++ src/cmds/versions/update.ts
@@ -4,7 +4,7 @@ import { Args, Flags } from '@oclif/core';
 import { Headers } from 'node-fetch';
 import prompts from 'prompts';
 
-import BaseCommand from '../../lib/baseCommandNew.js';
+import BaseCommand from '../../lib/baseCommand.js';
 import castStringOptToBool from '../../lib/castStringOptToBool.js';
 import { baseVersionFlags, keyFlag } from '../../lib/flags.js';
 import * as promptHandler from '../../lib/prompts.js';
